replace old interpreter. isolate context class
